@@ -17,10 +17,12 @@
 import java.util.*;
 import edu.kit.iti.checker.property.subchecker.lattice.qual.*;
 
-// :: error: initialization.fields.uninitialized
 public class NullnessDefaultTest {
+    // :: error: initialization.field.uninitialized
     @NonNull Object nonNullField;
+    // :: error: initialization.field.uninitialized
     @Nullable Object nullableField;
+    // :: error: initialization.field.uninitialized
     Object defaultField;
 
     public void foo() {
