@@ -1,33 +1,16 @@
-/* This file is part of the Property Checker.
- * Copyright (c) 2021 -- present. Property Checker developers.
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details.
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- */
+import org.checkerframework.checker.nullness.qual.*;
+
 public class Test {
-
-    Object field = new Object();
     
-    void foo() { }
-}
-
-class A {
-
-    public static void foo() {
-        Test test1 = new Test();
-        Test test2 = new Test();
-        test1.foo();
-        test2.foo();
+    public static @NonNull Object bar(@Nullable Object arg) {
+        Object local = null;
+        if (arg != null) {
+            local = arg;
+            arg = null;
+            local = arg;
+            // :: error: return.type.incompatible
+            return local;
+        }
+        return new Object();
     }
-    
 }
